spinning icon fix
