+++ assets/svg/refetch.js
@@ -1,0 +1,19 @@
+import * as React from "react"
+
+function SvgComponent(props) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 24 24"
+      fill="#fff"
+      width={16}
+      height={16}
+      {...props}
+    >
+      <path d="M0 0h24v24H0z" fill="none" />
+      <path d="M19 8l-4 4h3c0 3.31-2.69 6-6 6a5.87 5.87 0 01-2.8-.7l-1.46 1.46A7.93 7.93 0 0012 20c4.42 0 8-3.58 8-8h3l-4-4zM6 12c0-3.31 2.69-6 6-6 1.01 0 1.97.25 2.8.7l1.46-1.46A7.93 7.93 0 0012 4c-4.42 0-8 3.58-8 8H1l4 4 4-4H6z" />
+    </svg>
+  )
+}
+
+export default SvgComponent
