+++ components/DoubleStatsBox.tsx
@@ -6,13 +6,7 @@ import { getFormattedNumber } from 'utils/formatter';
 import InfoPopover from './InfoPopover';
 import { UseQueryResult } from 'react-query';
 
-import { SnxTooltip } from './common';
-
-import { IconButton, withStyles } from '@material-ui/core';
-
-import ErrorIcon from 'assets/svg/error';
-import RefetchIcon from 'assets/svg/refetch.svg';
-import { useTranslation } from 'react-i18next';
+import StatsTools from './StatsTools';
 
 type DoubleStatsBoxProps = {
 	title: string;
@@ -43,12 +37,7 @@ const DoubleStatsBox: FC<DoubleStatsBoxProps> = ({
 	queries = [],
 	infoData,
 }) => {
-	const { t } = useTranslation();
-
 	const allQueriesLoaded = !queries.find((q) => q.isLoading);
-	const hasQueryError = !!queries.find((q) => q.isError);
-
-	const refetch = () => queries.forEach((q) => q.refetch());
 
 	const formattedFirstMetric =
 		allQueriesLoaded && firstMetric != null
@@ -67,14 +56,7 @@ const DoubleStatsBox: FC<DoubleStatsBoxProps> = ({
 					<DoubleStatsBoxTitle>{title}</DoubleStatsBoxTitle>
 					{infoData != null ? <InfoPopover infoData={infoData} /> : null}
 				</TitleWrapper>
-				<InfoWrapper>
-					{hasQueryError && <WarningIcon />}
-					<SnxTooltip arrow title={t('refresh-tooltip') as string} placement="top">
-						<RefetchIconButton aria-label="refetch" onClick={refetch}>
-							<RefetchIcon />
-						</RefetchIconButton>
-					</SnxTooltip>
-				</InfoWrapper>
+				<StatsTools queries={queries} />
 			</HeaderWrapper>
 			<DoubleStatsBoxSubtitle>{subtitle}</DoubleStatsBoxSubtitle>
 			<DoubleStatsBoxMetricTitle>{firstMetricTitle}</DoubleStatsBoxMetricTitle>
@@ -109,25 +91,6 @@ const HeaderWrapper = styled.div`
 	margin-bottom: 15px;
 	margin-right: 10px;
 `;
-
-const InfoWrapper = styled.div`
-	display: flex;
-	justify-content: flex-end;
-	align-items: center;
-	margin-right: -10px;
-`;
-
-const WarningIcon = styled(ErrorIcon)`
-	fill: ${(props) => `${props.theme.colors.red}`};
-	margin-right: 10px;
-`;
-
-const RefetchIconButton = withStyles(() => ({
-	root: {
-		backgroundColor: '#312065',
-		padding: '6px',
-	},
-}))(IconButton);
 
 export const TitleWrapper = styled.div`
 	display: flex;
